// lib/types.ts
interface UserState {
  username: string | null;
  role: string | null;
  companyName: string | null;
  vendorType: string | null;
  isLoggedIn: boolean;
  login: (userData: {
    username: string;
    role: string;
    companyName: string | null;
    vendorType: string | null;
  }) => void;
  logout: () => void;
}

export const useAuthStore = create<UserState>()(
  persist(
    (set) => ({
      username: null,
      role: null,
      companyName: null,
      vendorType: null,
      isLoggedIn: false,

      login: (userData) =>
        set({
          username: userData.username,
          role: userData.role,
          companyName: userData.companyName,
          vendorType: userData.vendorType,
          isLoggedIn: true,
        }),

      logout: () =>
        set({
          username: null,
          role: null,
          companyName: null,
          vendorType: null,
          isLoggedIn: false,
        }),
    }),
    {
      name: 'auth-storage',
      storage: createJSONStorage(() => localStorage),
    }
  )
);

export type NullString = {
  String: string;
  Valid: boolean;
};
export type NullTime = {
  Time: string;
  Valid: boolean;
};
export type SqlNullString = { String: string; Valid: boolean };
export type SqlNullTime = { Time: string; Valid: boolean };

export type User = {
  id: number;
  username: string;
  role: string;
  companyName: NullString | null;
  vendorType: NullString | null;
};
export interface Vendor {
  id: number;
  companyName: string;
  vendorType: string;
  createdAt: string;
  updatedAt: string;
}
export interface Project {
  id: number;
  wbsNumber: string;
  projectName: string;
  vendorName: NullString | null;
  createdAt: string;
  updatedAt: string;
}


export interface BOM {
  id: number;
  bomCode: string;
  versionTag: string; 
  partReference: string;
  material: string;
  materialDescription: string;
  qty: number;
  createdAt: string;
  updatedAt: string;
}
// ...
export interface ActualPart {
  material: string;
  qty: number;
  views: string[];
}
// ...
export interface ProjectTrackingPayload {
  projectId: { Int64: number; Valid: boolean } | null;
  switchboardName: string;
  compartmentNumber: string;
  mechAssemblyBy: SqlNullString | null;
  wiringType: SqlNullString | null;
  wiringBy: SqlNullString | null;
  statusTest: "Waiting" | "Tested" | "Already Compared with BOM";
  testedBy: SqlNullString | null;
  dateTested: SqlNullTime | null;
  actualParts: ActualPart[] | null;
}